Fix tiny metal bug

+++ risc0/circuit/rv32im-m3-sys/cxx/hal/metal/hal.cpp
@@ -41,7 +41,7 @@ public:
   MetalBuffer(MTL::Buffer* buffer) : buffer(buffer) {}
   ~MetalBuffer() override { buffer->release(); }
   size_t size() override { return buffer->length(); }
-  void copyFromCpu(uint32_t offset, const void* data, size_t size) override {
+  void copyFromCpu(size_t offset, const void* data, size_t size) override {
     memcpy(reinterpret_cast<uint8_t*>(buffer->contents()) + offset, data, size);
   }
